Fixed: moved pacakge models form package router
